merge: ci fix, update circe
Overall, it seems that the old circe will cause compile error in action
environment. But I don't know why and it cannot be reporduced in my
computer.

I also added Xss args, but it should not work due to I cannot reproduce
the issue on my computer with this arg.

I also change the CI Java from termurin to liberica, which is the Java
that used in my computer. But it also makes no change to the issue.

commit f7496bf26968c4def377bbb22d88bef8b61c9aa8
Author: Eyre_S <sukazyo@outlook.com>
Date:   Fri Jun 27 23:05:48 2025 +0800

    ci fix 2 - update circe & add Xss

commit 4f39a15b0b0726665eec52be2b0807e90bdf50d9
Author: Eyre_S <sukazyo@outlook.com>
Date:   Fri Jun 27 23:01:17 2025 +0800

    ci fix test 1